+++ classes/Reports/ipnlog.class.php
@@ -237,16 +237,7 @@ class ipnlog extends \Shop\Report
                 }
             }
             if ($ipn) {
-                $T->set_var('ipn_data', true);
-                $T->set_block('report', 'rawBlock', 'Raw');
-                foreach ($ipn as $name => $value) {
-                    $T->set_var(array(
-                        'name'  => $name,
-                        'value' => htmlspecialchars($value),
-                        //'value' => $value,
-                    ) );
-                    $T->parse('Raw', 'rawBlock', true);
-                }
+                $T->set_var('ipn_data', print_r($ipn, true));
             }
             $retval = $T->parse('output', 'report');
         }